feat(board): show walking distance  (#1444)
* feat(board): show walking distance

* chore(edit): rename

* feat(location): update walk distance

* chore(address): walking distance settings

* fix(address): remove location

* styling(address): time forma

* chore(address): toast on save

* chore(address): remove unused

* chore(address): remove duplicate

* chore(address): rename parameter

* chore(address): walking distance component

* chore(address): helper function

* chore(address): refactor

* chore(address): improvements

* chore(address): simplify function

* chore(address): update imports

* chore(address): improvements

* chore(address): return early if distance undefined

* chore(address): fallback

* chore(address): access control

* chore(address): refactor

* chore(address): return early

+++ next-tavla/app/(admin)/edit/[id]/components/MetaSettings/actions.ts
@@ -8,7 +8,8 @@ import { firestore } from 'firebase-admin'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 import { TFontSize, TLocation } from 'types/meta'
-import { TBoardID } from 'types/settings'
+import { TBoard, TBoardID } from 'types/settings'
+import { getBoard, getWalkingDistanceTile } from '../../actions'
 
 initializeAdminApp()
 
@@ -46,12 +47,23 @@ export async function saveLocation(bid: TBoardID, location?: TLocation) {
     const access = hasBoardEditorAccess(bid)
     if (!access) return redirect('/')
 
+    const board = await getBoard(bid)
+    if (!board) return getFormFeedbackForError('board/not-found')
     await firestore()
         .collection('boards')
         .doc(bid)
         .update({
+            tiles: await getTilesWithDistance(board, location),
             'meta.location': location ?? firestore.FieldValue.delete(),
             'meta.dateModified': Date.now(),
         })
     revalidatePath(`/edit/${bid}`)
+}
+
+async function getTilesWithDistance(board: TBoard, location?: TLocation) {
+    return await Promise.all(
+        board.tiles.map(async (tile) => {
+            return await getWalkingDistanceTile(tile, location)
+        }),
+    )
 }
